assert decimal type'
